requirements.txt
aiosmtplib==2.0.2
alabaster==0.7.13
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
Babel==2.13.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.36.0
cryptography==41.0.4
dnspython==2.4.2
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.0.0.post2
Faker==19.10.0
fastapi==0.103.2
fastapi-limiter==0.1.5
fastapi-mail==1.4.1
greenlet==3.0.0
h11==0.14.0
hiredis==2.2.3
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.9
packaging==23.2
passlib==1.7.4
pika==1.3.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pypng==0.20220715.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
